Implement no_std runtime
