+++ Cargo.toml
@@ -19,13 +19,16 @@ runtime_asyncstd = [
 runtime_tokio = [
     "tokio",
 ]
-runtime_nostd = []
+runtime_nostd = [
+    "pin-utils",
+]
 
 [dependencies]
 bastion-executor = { version = "0.3", optional = true, features = [] }
 async-std = { version = "1.3", optional = true, features = ["unstable"] }
 tokio = { version = "0.2", optional = true, features = ["rt-core", "blocking", "rt-util"] }
 lightproc = { version = "0.3.4", optional = true }
+pin-utils = { version = "0.1.0", optional = true }
 
 [dev-dependencies.tokio]
 version = "0.2"
